app/backend/requirements.txt
flask
flask-login
flask-bcrypt
flask-cors
pyopenssl
pymongo
python-dotenv
gunicorn